Move notification code out of CommCareApplication
