+++ app/src/org/commcare/activities/CommCareSetupActivity.java
@@ -621,7 +621,7 @@ public class CommCareSetupActivity extends CommCareActivity<CommCareSetupActivit
     private void fail(NotificationMessage notificationMessage, boolean showAsPinnedNotifcation) {
         String message;
         if (showAsPinnedNotifcation) {
-            CommCareApplication.instance().reportNotificationMessage(notificationMessage);
+            CommCareApplication.noficationManager().reportNotificationMessage(notificationMessage);
             message = Localization.get("notification.for.details.wrapper",
                     new String[]{notificationMessage.getTitle()});
         } else {
@@ -663,7 +663,7 @@ public class CommCareSetupActivity extends CommCareActivity<CommCareSetupActivit
 
     @Override
     public void reportSuccess(boolean newAppInstalled) {
-        CommCareApplication.instance().clearNotifications("install_update");
+        CommCareApplication.noficationManager().clearNotifications("install_update");
 
         if (newAppInstalled) {
             GoogleAnalyticsUtils.reportAppInstall(lastInstallMode);
