+++ app/src/org/commcare/activities/MessageActivity.java
@@ -34,7 +34,7 @@ public class MessageActivity extends ListActivity {
         if (savedInstanceState != null && savedInstanceState.containsKey(KEY_MESSAGES)) {
             messages = savedInstanceState.getParcelableArrayList(KEY_MESSAGES);
         } else {
-            messages = CommCareApplication.instance().purgeNotifications();
+            messages = CommCareApplication.noficationManager().purgeNotifications();
         }
     }
