+++ app/src/org/commcare/adapters/EntityListAdapter.java
@@ -479,6 +479,6 @@ public class EntityListAdapter implements ListAdapter, EntitySortNotificationInt
 
     @Override
     public void notifyBadfilter(String[] args) {
-        CommCareApplication.instance().reportNotificationMessage(NotificationMessageFactory.message(NotificationMessageFactory.StockMessages.Bad_Case_Filter, args));
+        CommCareApplication.noficationManager().reportNotificationMessage(NotificationMessageFactory.message(NotificationMessageFactory.StockMessages.Bad_Case_Filter, args));
     }
 }
